fix(intraday): verify every ticker, Eastern loader bounds, session dates
Four review findings on 68ee4ea:

- classify_intraday_ts_convention.sql divided BIGINT volume sums, so a
  0.5 ratio truncated to 0 (true-UTC) and 1.8 to 1 (mixed). Cast to
  numeric. On production SPY/QQQ/IWM: 35 ET-as-UTC and 70 mixed
  sessions, against 34 and 71 before.
- That classifier covers SPY/QQQ/IWM only, and reading every row of
  every ticker does not fit a statement timeout. Add a read-only
  --verify-months mode: per ticker-month, read the window and require
  stored_intraday_to_eastern to read every row as a true instant with
  nothing dropped; VERIFY-FAIL lines and exit 1 otherwise. Run it over
  a list regenerated from the table, so a month the manifest omitted
  is checked too. About 0.6 s per SPY-sized month, no vendor calls.
- DataLoader bound Eastern date strings straight to TIMESTAMPTZ, where
  they read as UTC: an EST request for D pulled in D-1's 19:00-20:00 ET
  spill and lost D's own. The raw window now runs from start (as a
  label) to end converted with eastern_index_to_utc, and the converted
  index is cut back to the Eastern bounds.
- The Charts date list dropped raw hours 0-3, losing a session whose
  only retained bars were the next-day spill. Date every row by its
  session, (raw - 4 h)::date. Production IWM: 2,941 dates either way,
  0 weekends, 1.60 s against 2.61 s.

CLAUDE.md 3.9 now names --verify-months as the migration's check.

+++ gcp/queries/classify_intraday_ts_convention.sql
@@ -25,6 +25,13 @@
 -- intraday bars by ~190 bp on 2024 dates under BOTH readings (consistent
 -- with dividend adjustment), so it cannot arbitrate.
 --
+-- Scope: SPY, QQQ, IWM only. Reading every row of every ticker does not fit a
+-- statement timeout (1/16 of the catch-all partition alone timed out at
+-- 540 s). The all-ticker check after the re-framing migration is
+--   python -m gcp.fetchers.fetch_alphavantage_intraday --verify-months <list>
+-- with the list regenerated from the table by list_intraday_ticker_months.sql,
+-- so a month the migration manifest omitted is checked too (Codex P1 on #1185).
+--
 -- Usage (read-only):
 --   ./scripts/db_query_cr.sh -f gcp/queries/classify_intraday_ts_convention.sql --timeout 300
 WITH per_session AS (
@@ -54,8 +61,10 @@ SELECT ticker, d AS session_date, v_true_instant, v_et_label,
        CASE
          WHEN coalesce(v_true_instant, 0) + coalesce(v_et_label, 0) = 0 THEN 'no-open-bars'
          WHEN coalesce(v_true_instant, 0) = 0 THEN 'ET-as-UTC'
-         WHEN coalesce(v_et_label, 0) / v_true_instant > 1.5 THEN 'ET-as-UTC'
-         WHEN coalesce(v_et_label, 0) / v_true_instant < 0.2 THEN 'true-UTC'
+         -- numeric: volume is BIGINT, and integer division turned a 0.5
+         -- ratio into 0 (true-UTC) and 1.8 into 1 (mixed) (Codex P1 on #1185).
+         WHEN coalesce(v_et_label, 0)::numeric / v_true_instant > 1.5 THEN 'ET-as-UTC'
+         WHEN coalesce(v_et_label, 0)::numeric / v_true_instant < 0.2 THEN 'true-UTC'
          ELSE 'mixed'
        END AS convention
 FROM per_session
